refactore store.js - modularize into separate files per store prop
