+++ packages/vuetron-app/src/store/state.js
@@ -1,0 +1,388 @@
+const state = {
+  clientState: {},
+  events: [],
+  subscriptions: {},
+  domTree: {
+    "children": [
+      {
+        "children": [
+          {
+            "children": [
+              {
+                "children": [
+                  {
+                    "children": [
+
+                    ],
+                    "id": 53,
+                    "name": "DiscogsEntityType"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 50,
+                    "name": "DiscogsImageFormatType"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 52,
+                    "name": "DiscogsPaginable"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 54,
+                    "name": "DiscogsSearch"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 51,
+                    "name": "DiscogsSortInformation"
+                  }
+                ],
+                "id": 49,
+                "name": "Query"
+              },
+              {
+                "children": [
+                  {
+                    "children": [
+
+                    ],
+                    "id": 14,
+                    "name": "DiscogsArtist"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 13,
+                    "name": "DiscogsArtistRelease"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 46,
+                    "name": "DiscogsArtistReleases"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 23,
+                    "name": "DiscogsArtistSortType"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 27,
+                    "name": "DiscogsCommunity"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 26,
+                    "name": "DiscogsCommunityInfo"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 15,
+                    "name": "DiscogsCommunityReleaseRating"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 28,
+                    "name": "DiscogsEntity"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 29,
+                    "name": "DiscogsFormat"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 22,
+                    "name": "DiscogsGroupOrBandMember"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 30,
+                    "name": "DiscogsIdentifier"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 16,
+                    "name": "DiscogsIdentity"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 40,
+                    "name": "DiscogsImage"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 31,
+                    "name": "DiscogsImageType"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 32,
+                    "name": "DiscogsLabel"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 17,
+                    "name": "DiscogsLabelRelease"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 48,
+                    "name": "DiscogsLabelReleases"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 24,
+                    "name": "DiscogsMaster"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 19,
+                    "name": "DiscogsPaginableResults`1"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 33,
+                    "name": "DiscogsPaginedResult"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 34,
+                    "name": "DiscogsPaginedUrls"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 25,
+                    "name": "DiscogsRating"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 41,
+                    "name": "DiscogsRelease"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 39,
+                    "name": "DiscogsReleaseArtist"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 35,
+                    "name": "DiscogsReleaseBase"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 18,
+                    "name": "DiscogsReleaseLabel"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 20,
+                    "name": "DiscogsReleaseRating"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 44,
+                    "name": "DiscogsReleaseVersion"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 47,
+                    "name": "DiscogsReleaseVersions"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 36,
+                    "name": "DiscogsSearchResult"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 42,
+                    "name": "DiscogsSearchResults"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 37,
+                    "name": "DiscogsSimplifiedLabel"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 21,
+                    "name": "DiscogsSortOrderType"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 38,
+                    "name": "DiscogsTrack"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 43,
+                    "name": "DiscogsUser"
+                  },
+                  {
+                    "children": [
+
+                    ],
+                    "id": 45,
+                    "name": "DiscogsVideo"
+                  }
+                ],
+                "id": 12,
+                "name": "Result"
+              }
+            ],
+            "id": 11,
+            "name": "Data"
+          },
+          {
+            "children": [
+              {
+                "children": [
+
+                ],
+                "id": 9,
+                "name": "DiscogsWebClient"
+              },
+              {
+                "children": [
+
+                ],
+                "id": 10,
+                "name": "IDiscogsWebClient"
+              }
+            ],
+            "id": 8,
+            "name": "Internal"
+          },
+          {
+            "children": [
+
+            ],
+            "id": 2,
+            "name": "DiscogsAuthentifierClient"
+          },
+          {
+            "children": [
+
+            ],
+            "id": 3,
+            "name": "DiscogsClient"
+          },
+          {
+            "children": [
+
+            ],
+            "id": 4,
+            "name": "DiscogsException"
+          },
+          {
+            "children": [
+
+            ],
+            "id": 5,
+            "name": "IDiscogsDataBaseClient"
+          },
+          {
+            "children": [
+
+            ],
+            "id": 6,
+            "name": "IDiscogsReleaseRatingClient"
+          },
+          {
+            "children": [
+
+            ],
+            "id": 7,
+            "name": "IDiscogsUserIdentityClient"
+          }
+        ],
+        "id": 1,
+        "name": "DiscogsClient"
+      }
+    ],
+    "id": 0,
+    "name": "DiscogsClient"
+  },
+  // VUETRON APP STATE PROPS
+  displayNavbar: true
+};
+
+export default state;
